fix: Handle when a cell has no data using `grid`
When a chart is using the `grid` feature of ECharts for layout and a cell has no data, there is no `grid` that gets output, making the references to `gridIndex`, ‘xAxisIndex’ and ‘yAxisIndex’ wrong. This passes the render index rather than the index of the data series to allow the correct index to be used.