refactor: Revert to `xAxis`, `yAxis` for some variables
Styles are specific to the orientation of the axis, not their content. So much of the calculation for the axes needs to be done based on the rendered axis, not the logical axis (i.e., `x/y` not `category/value`).

In the next commit I’ll add the little bit of code that needs to vary based on the orientation.