feat: Add `legendOrientation` argument
Make the legend orientation independent from where it is positioned. So we can have a vertically oriented legend positioned underneath the chart.

Also fix the documentation for `computeMaxTextMetrics` to make it clearer that it is not taking the height of all of the text, but rather the maximum height required to render any of the text strings.