feat: Add `categoryAxisSort` argument to bar chart
By default the categories are rendered in the order they exist in the first series (or if not in that series, in subsequent series). This causes problems when the first series doesn’t have all of the categories, or doesn’t have them in the correct order.

This new prop solves the issue by allowing a sort order be specified when the default order doesn’t make sense.