feat: Use data objects for data instead of values
To make it easier to grab the full data object in tooltips, pass the entire data object into `data` instead of only the values.

Note that these still must be in the same order as the categories, with any items that don’t match a category as `undefined`, which is what `getSeriesData` does.