fix: Left margin on Y axis
The width of the Y axis was being calculated using values derived from the data, not the actual labels that would be rendered on the Y axis. This was made worse when the wrong values were used for grouped variants, especially if one of the values had lots of decimal places that wouldn’t be rendered when rounded on the axis.

This commit does a few things:

- Refactors the calculation of the values & stats into `computeValuesInfo`
- Fixes the calculation of values for the grouped variant in `computeValuesInfo`
- Uses the actual scale from eCharts to calculate the tick labels
- Calculates the width based on these tick labels rather than the approximate value labels from before