fix: Bottom margin on X axis
This is the companion commit to the last commit which fixed the left margin on the Y axis.

This commit itself fixes only the edge case where the `categoryAxisFormatter` formats to a shorter label than the original category (i.e., if a category is long enough, it will wrap and require multiple lines to render, which the formatter then removes, but the layout algorithm already accounted for those extra lines).

However, this commit sets up the category axis to be formatted correctly when using a `time` axis (since a Date, if not formatted by a formatter, is quite long and will wrap to multiple lines).

This commit does similar things to the last commit:

- Refactors the calculation of the categories & stats into `computeCategoryInfo`
- Uses the actual scale from eCharts to calculate the tick labels for the category axis
- Calculates the height of the category axis based on these tick labels