feat: Add `time` category axis type
Add the ability to specify the category axis as a `time` axis.

Two key differences between the `time` and `category` axis types are:

- Lines and areas are drawn through unspecified datetime points in the `time` axis whereas the `category` axis only draws points explicitly where the data indicates and then only connects points that are next to one another
- When the `time` axis is zoomed into, the labels will adjust based on the axis extent, whereas the labels for the `category` axis will remain fixed.

Adding `time` as an axis type into this chart type is a bit of a hack, but it’s easier than re-creating an entire new timeseries chart type that has almost the same implementation.

To implement the `time` axis type, a couple of things are happening:

- `computeCategoryAxisLabels` is using a time scale to compute the labels using the default time label formatter
- The category axis is being configured as a `time` type (and with `data` removed, since that only applies for the `category` type)
- The data passed into `series` is enhanced to use tuples for the `value` field, since when using the `time` axis type, the `name` property of the data objects is ignored (the tuples are [x, y], so are flipped when the chart is being rendered as a horizontal chart)
- The tooltip is extracting the real value from the tuple so the tooltips render as expected when using a `time` axis
- The categories are being sorted from oldest to newest when calling `getCategories`—this may no longer be strictly necessary, but is nice if we ever need the raw category data

Finally, `setOption` is being called on the chart when the chart is first created. This is to initialize the chart model so the locale defined within this model can be used when generating the labels for the time axis.