fix: Label cut off on value axis on horizontal chart
If the X axis is a value axis, then the label is placed at the end of the axis; otherwise it is placed in the middle of each band. So this problem appears to only occur when the X axis is a value axis, i.e., when the chart has a `horizontal` orientation.