fix: Incorrect axis metrics by using real ticks
The width and height of the axes wasn’t always correct because we weren’t taking into account label placement on the axes. This switches to calculating the full tick info rather than just the labels and then using this to adjust the height and width overflow for the axes.

It also fixes a layout issue for the time axis where the type of the tick was being embedded in the tick label and potentially causing the wrong width of the label to be calculated.