feat: Make chart resize under more layout situations
The chart wasn’t resizing easily for certain layout conditions, such as when put within a grid. Now if you set the `@height` to `100%`, it should resize to the full height of the container element in all layouts.

If the parent element doesn’t have a height because it hasn’t been configured to have one (e.g., its parent isn’t using `flex` or `grid` layout and it doesn’t have an explicit `height` or `min-height` set), then the chart will just default to 600px wide by 400px tall.